requirements.txt
selene>=2.0.0rc7
pytest~=8.3.4
openpyxl~=3.2.0b1
PyPDF2~=3.0.1